Proper Thresholds for BB and HCal + 99.5% Energies
Finished implementing the proper thresholds for BB preshower + shower energies as well as HCal energy threshold. Can calculate as % or max energy in an event or as a hard coded value. Now finds event and hit energies with 99.5% of other event or hit energies below it and outputs this value.