HCal NIM Paper Update
Added in a description of the signal flow throughout the front-end and DAQ side. I think it still needs a fair amount of work to format this better. Maybe I don't need to do more than make a block diagram? Not wasted since it can go in the HCal user's manual regardless.