Creates a GUI with clickable PMT buttons which give data.
This script creates a GUI that has buttons in the configuration of HCal's PMTs. When a PMT button is clicked, the terminal displays connections data such as what pre-amp, fADC, and HV channel are associated with the given PMT. This will hopefully be expanded to cover the FE and DAQ racks for HCal, but that's a much bigger project.