Finished Geometry of Front-End Electronics
Geometry of front-end electronics is now all coded. Still need to bind the connections functions to their buttons, but this requires the super connections dictionary. Before making this dictionary I would like to make the DAQ side electronics geometry. Also got a scroll bar working, but it did mess up expanding the container frame automatically so width had to be hard coded. Not sure this is worth figuring out.