Finished Adding Connections to Buttons and Row/Col
I added the connection info to most of the buttons now for HCal, the front-end, and the DAQ. Need to do things like TI and NIM crates still. Also added row-col infor after listing PMT number (1-288) and put that on the HCal GUI buttons too. Also shaded some of the HCal GUI buttons to show where the clusters are and make it a bit easier to read. Lots of improvements to be made, but I think this is now a functional version of the initial idea.