Started HCal Event Display GUI
Trying to make a new HCal event display with tkinter and matplotlib (don't think I can get ROOT to plot on a tkinter canvas). I have a single fADC channel plotting on a tkinter canvas, but it keeps adding more canvases instead of overwriting them. Need to figure out a way to delete the canvas object if it's present.