Implemented Plotting All fADC Waveforms
Fixed many of the earlier problems. I was defining the canvases twice (once in the main body and once in the function) which was causing them to plot repeatedly (fixed it by not redefining them, but just changing the histos plotted on them like I should). Figured out how to make a canvas clickable so that it pops up a new larger window with a single fADC waveform (very cool). Got all 288 fADC waveforms plotting on one window, but had to add a scrollbar to see everything. Added next and previous buttons to skip ahead or back one event (need to add protections for <0 and >max event still). It seems a bit slow to plot, but I'll try to see if I can do this more efficiently (could just be my laptop too). Need to add color for TDC hits or above an ADC threshold. Need to add other types of plots with either their own buttons or with a radio button swapping between them. Could get pedestal from first 4 events (or better yet rootfile with updated SBS-Offline) to define the baseline of each plot and do a ped subtraction if annoyed by the different pedestal heights. This would make the peaks more comparable to one another visually. Need to add better comments and clean up old code before giving this to others.